+++ src/main/java/com/Joysbright/digiBank/dtos/request/UserTransferRequest.java
@@ -3,6 +3,7 @@ package com.Joysbright.digiBank.dtos.request;
 import lombok.*;
 
 import javax.swing.*;
+import java.math.BigDecimal;
 
 @Getter
 @Setter
@@ -13,6 +14,6 @@ public class UserTransferRequest {
     private  String senderAccountNo;
     private String senderAccountPassword;
     private String ReceiverAccountNo;
-    private double amountToSend;
+    private BigDecimal amountToSend;
     private String narration;
 }
